% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ep_wp_variables.R
\name{add_air_yac_wp_variables}
\alias{add_air_yac_wp_variables}
\title{Calculate and add the air and yac win probability variables to include in 
a `nflscrapR` play-by-play data frame}
\usage{
add_air_yac_wp_variables(pbp_data)
}
\arguments{
\item{pbp_data}{Data frame with all of the necessary columns used to estimate
and include the air and yac wpa}
}
\value{
The input data frame with additional columns included for the 
air WPA (air_wpa), yac WPA (yac_wpa), and cumulative totals for home and away
teams (total_home_air_wpa, total_home_yac_wpa, etc.).
}
\description{
Given a `nflscrapR` play-by-play data frame, calculate the air and yac WPA 
for passing playse using the `nflscrapR` win probability model and include append
the column to the input data frame. See here for an explanation of the 
model methodology: \url{https://arxiv.org/abs/1802.00998}. Source code for 
fitting the model is located here \url{https://github.com/ryurko/nflscrapR-models/blob/master/R/init_models/init_ep_fg_models.R}.
}
