% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ep_wp_variables.R
\name{add_wp_variables}
\alias{add_wp_variables}
\title{Calculate and add the win probability variables to include in a `nflscrapR` 
play-by-play data frame}
\usage{
add_wp_variables(pbp_data)
}
\arguments{
\item{pbp_data}{Data frame with all of the necessary columns used to estimate
the win probability for a play.}
}
\value{
The input data frame with additional columns included for the 
win probability (wp), win probability added (wpa), and respective  
win probability for both home and away teams (home_wp, away_wp).
}
\description{
Given a `nflscrapR` play-by-play data frame, calculate the win probability
for a play using the `nflscrapR` win probability model and include 
the columns to the input data frame. See here for an explanation of the 
model methodology: \url{https://arxiv.org/abs/1802.00998}. Source code for 
fitting the model is located here \url{https://github.com/ryurko/nflscrapR-models/blob/master/R/init_models/init_ep_fg_models.R}.
}
