% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlayerGameFunctions.R
\name{agg_player_season}
\alias{agg_player_season}
\title{Detailed Player Aggregate Season Statistics}
\usage{
agg_player_season(Season, Weeks = 16)
}
\arguments{
\item{Season}{(numeric) A 4-digit year associated with a given season}
}
\value{
Returns a dataframe with a single line for each player aggregating 
their total season statistics
}
\description{
This function outputs a dataframe with the season statistics for
each player who recorded atleast one measured statistic in any game throughout
the specified season.  This function gives one line per player with the 
following statistics: Passing, Rushing, Receiving, Kick Return,
Punt Return, Fumbles, and Defense
}
\details{
This function calls season_player_game and then aggregates 
across an entire season to gather season totals and season max statistics
}
\examples{
# Returns the Season-Total Statistics for Each Player in the 2015 Season
agg_player_season(2015)
}
