% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rosters.R
\name{build_name_abbr}
\alias{build_name_abbr}
\title{Build formatted player name from full player name}
\usage{
build_name_abbr(value)
}
\description{
This sub-function, called in the get_season_rosters function,
takes the full name of each player and formats it into the first initial of 
their first name and last initial of their last name.
}
