% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rosters.R
\name{build_url}
\alias{build_url}
\title{Build URL to scrape player season stat pages}
\usage{
build_url(position, season, page = 1, type)
}
\arguments{
\item{position}{(character string) Specifies a player position page for the URL}

\item{season}{4-digit year associated with a given NFL season}

\item{page}{1-digit page number to look into}

\item{type}{A three character string specifying the season type: pre, post, or reg}
}
\description{
This is a sub-function for the get_season_rosters function.
}
