% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameandRosterFunctions.R
\name{buildNameAbbr}
\alias{buildNameAbbr}
\title{Build formatted player name from full player name}
\usage{
buildNameAbbr(value)
}
\description{
This sub-function, called in the season_rosters function,
takes the full name of each player and formats it into the first initial of 
their first name and last initial of their last name.
}
