% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameandRosterFunctions.R
\name{buildURL}
\alias{buildURL}
\title{Building URL to scrape player season stat pages}
\usage{
buildURL(position, season = 2016, page = 1, type = c("REG", "POST",
  "PRE"))
}
\arguments{
\item{position:}{(character string) Specifies a player position page for the URL}

\item{season:}{4-digit year associated with a given NFL season}

\item{page:}{1-digit page number to look into}

\item{type:}{A three character string specifying the season type}
}
\description{
This is a sub-function for the season_rosters
function.
}
