% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_wp_calculator.R
\name{calculate_win_probability}
\alias{calculate_win_probability}
\title{Compute win probability for the provided play-by-play dataset.}
\usage{
calculate_win_probability(pbp_data, half_seconds_remaining,
  game_seconds_remaining, score_differential, quarter, posteam_timeouts_pre,
  oppteam_timeouts_pre, ep)
}
\arguments{
\item{pbp_data}{Play-by-play dataset to estimate expected points for.}

\item{half_seconds_remaining}{String denoting the name of the column of the
pbp_data containing the number of seconds remaining in the current half.}

\item{game_seconds_remaining}{String denoting the name of the column of the
pbp_data containing the number of seconds remaining in the game.}

\item{score_differential}{String denoting the name of the column of the pbp_data
containing the score differential with respect to the possession team.}

\item{quarter}{String denoting the name of the column of the pbp_data containing
the quarter of the play (anything above 4 is considered overtime).}

\item{posteam_timeouts_pre}{String denoting the name of the column of the pbp_data 
containing the number of timeouts remaining for the team with possession.}

\item{oppteam_timeouts_pre}{String denoting the name of the column of the pbp_data
containing the number of timeouts remaining for the opposing team.}

\item{ep}{String denoting the name of the column of the pbp_data containing
the expected points with respect to the possession team for the play.}
}
\value{
The original pbp_data with a column named wp containing the win probability
with respect to the team with possession.
}
\description{
Given a dataset of plays and the necessary variables, this function returns
the original dataset with the win probability from the nflscrapR model.
}
