% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_games_and_urls.R
\name{create_game_json_url}
\alias{create_game_json_url}
\title{Create the url with the location of NFL game JSON data}
\usage{
create_game_json_url(game_id)
}
\arguments{
\item{game_id}{(character or numeric) A 10 digit game id associated with a 
given NFL game.}
}
\value{
String url where the game JSON data for the given game can be found.
}
\description{
This function pastes creates the JSON url for the provided game id (first 
checking that it is from the 2009 season or later). The returned url is used
in the various play-by-play functions to access data from the NFL's API.
}
\examples{
# Create the JSON url for the first game of the 2017 season:
create_game_json_url(2017090700) 
}
