% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlayByPlayBoxScore.R
\name{drive_summary}
\alias{drive_summary}
\title{Drive Summary and Results}
\usage{
drive_summary(URLString)
}
\arguments{
\item{URLString:}{A string with the URL location of a single NFL
game JSON API}
}
\value{
A dataframe that has the summary statistics for each drive
     final output includes first downs, drive result, penalty yards,
     of plays, time of possession, quarter at the start of the drive,
     Time at Start of Drive, yardline at start of drive,
     team with possession at start, end of drive quarter, end of drive time,
     end of drive Yard line, end of drive team with possession
}
\description{
This function outputs the end result of each drive of a given
 game
}
\examples{
# Parsed drive Summarize of final game in 2015 NFL Season
nfl2015.finalregseasongame <- "http://www.nfl.com/liveupdate/game-center/2016010310/2016010310_gtd.json"
drive_summary(nfl2015.finalregseasongame)
}

