% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScrapingFunctionsGameURLs.R
\name{extracting_gameids}
\alias{extracting_gameids}
\title{Extract GameIDs for each game in a given NFL season}
\usage{
extracting_gameids(Season, playoffs = FALSE)
}
\arguments{
\item{Season}{(numeric) A numeric 4-digit year associated with an NFL season}

\item{playoffs}{(boolean) TRUE if you want the playoffs gameIDs for the specified 
season.  FALSE (default) if you want the game IDs from the regular season}
}
\value{
A vector of NFL GameIDs from the specified season
}
\description{
This function is a helper for the Proper.PBP.URL.Formatting function.
The outputs of extracting_gameids is used in conjunction with Proper.PBP.URL.Formatting
to create the URLs for each NFL game in a specified season
}
\examples{
# Scraping all game IDs from 2010 Season
extracting_gameids(2010) 
}
