% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rosters.R
\name{find_page_player_birthdate}
\alias{find_page_player_birthdate}
\title{Find the birthdate for each player on the provided page.}
\usage{
find_page_player_birthdate(value)
}
\description{
Find the birthdate for each player on the provided page.
}
