% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rosters.R
\name{find_page_player_id}
\alias{find_page_player_id}
\title{Find the GSIS ID for each player on the provided page.}
\usage{
find_page_player_id(value)
}
\description{
Find the GSIS ID for each player on the provided page.
}
