% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameandRosterFunctions.R
\name{findPagePlayerID}
\alias{findPagePlayerID}
\title{Find the GSIS ID for each player on the provided page.}
\usage{
findPagePlayerID(value)
}
\description{
Find the GSIS ID for each player on the provided page.
}
