% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlayByPlayBoxScore.R
\name{game_play_by_play}
\alias{game_play_by_play}
\title{Parsed Descriptive Play-by-Play Function for a Single Game}
\usage{
game_play_by_play(URLString)
}
\arguments{
\item{URLString}{(character) A URL string with the location of a single
NFL game JSON API}
}
\value{
A dataframe with 50 columns specifying various statistics associated
with each play of the specified NFL game.
}
\description{
This function intakes the JSON play-by-play data of a single
 game and parses the play description column individual variables allowing
 the user to segment the game in a variety of different ways allowing for
 in-depth analysis.
}
\details{
Through simple list manipulation using the dp.call function and
rbind this function creates a 10 column dataframe with basic information
 the NFL JSON API.  These columns include the following:
\itemize{
 \item{"Drive"}
 \item{"sp"} - Whether the play resulted in a score (any kind of score)
 \item{"qrt"}
 \item{"down"}
 \item{"time"} - Time at start of play
 \item{"yrdln"} - Between 0 and 50
 \item{"ydstogo"} - For a first down
 \item{"ydsnet"}
 \item{"posteam"} - The offensive team
 \item{"desc"} - A detailed description of what occured during the play
}

Through string manipulation and parsing of the desc column, 40 columns were
added to the original dataframe allowing the user to have a detailed
understanding of the events of each play.  The added variables are specified
below:
\itemize{
 \item{"Date", "GameID"}
 \item{"DefensiveTeam", "SideofField"}
 \item{"GoalToGo", "FirstDown", "PlayAttempted"}
 \item{"Yards.Gained", "TimeUnder"}
 \item{"ExPointResult", "TwoPointConv", "FieldGoalResult",
       "FieldGoalDistance", "TouchDown}
 \item{"PlayType"}
 \item{"Passer", "PassAttempt", "PassOutcome", "PassLength", "PassLocation",
       "InterceptionThrown"}
 \item{"Rusher", "RushAttempt", "RunLocation", "RunGap"}
 \item{"Receiver", "Reception"}
 \item{"Tackler1", "Tackler2", "Fumble", "Sack"}
 \item{"Accepted.Penalty", "PenalizedTeam", "Penalty Type",
       "Penalized Player", "Penalty.Yards"}
 \item{"PosTeamScore", "DefTeamScore", "ScoreDiff", "AbsScoreDiff"}
 }
}
\examples{
# Parsed play-by-play of the final game in the 2015 NFL season
nfl2015.finalregseasongame <- "http://www.nfl.com/liveupdate/game-center/2016010310/2016010310_gtd.json"
Game_PBP(nfl2015.finalregseasongame)
}

