% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rosters.R
\name{get_birthdate}
\alias{get_birthdate}
\title{For a player's href, get their birthdate from their personal url.}
\usage{
get_birthdate(value)
}
\description{
For a player's href, get their birthdate from their personal url.
}
