% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rosters.R
\name{get_gsis_id}
\alias{get_gsis_id}
\title{For a player's href, get their GSIS ID from their personal url.}
\usage{
get_gsis_id(value)
}
\description{
For a player's href, get their GSIS ID from their personal url.
}
