% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rosters.R
\name{get_page_numbers}
\alias{get_page_numbers}
\title{Get number of player position pages}
\usage{
get_page_numbers(value)
}
\description{
For each position, this function extracts the number of pages 
there are to scrape. This is a sub-function for the get_season_rosters function
}
