% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rosters.R
\name{get_players}
\alias{get_players}
\title{Scrape player names and positions}
\usage{
get_players(position, season, type)
}
\description{
This sub-function, calls build_name_abbr and get_page_numbers to
scrape player positions by season.
}
