% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rosters.R
\name{get_season_rosters}
\alias{get_season_rosters}
\title{Get team roster for a single season}
\usage{
get_season_rosters(season, teams, type = "reg", positions = c("QUARTERBACK",
  "RUNNING_BACK", "WIDE_RECEIVER", "TIGHT_END"))
}
\arguments{
\item{season}{A 4-digit year associated with a given NFL season}

\item{teams}{A string vector containing the abbreviations for NFL Team(s)}

\item{type}{String indicating the type of game, must either be: "pre", 
"reg", or "post".}

\item{positions}{A string vector containing the abbreviations for NFL position(s).
Can be any of the following:
\itemize{
\item{"QUARTERBACK"} (in the default list)
\item{"RUNNING_BACK"} (in the default list)
\item{"WIDE_RECEIVER"} (in the default list)
\item{"TIGHT_END"} (in the default list)
\item{"DEFENSIVE_LINEMAN"} 
\item{"LINEBACKER"}
\item{"DEFENSIVE_BACK"}
\item{"KICKOFF_KICKER"}
\item{"KICK_RETURNER"}
\item{"PUNTER"}
\item{"PUNT_RETURNER"}
\item{"FIELD_GOAL_KICKER"}
}}
}
\value{
A dataset with columns associated with season/year, full player name,
team initial, position, and formated player name.
}
\description{
Given a year and team abbreviation, return a dataset with each
player who has played the specified team and has recorded a measurable statistic.
}
\details{
To find team associated abbrevations use the \code{\link{nflteams}} dataset 
stored in this package!
}
\examples{
# Roster for Steelers in 2018
get_season_rosters(2018, teams = "PIT") 
}
