% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameandRosterFunctions.R
\name{getGSISID}
\alias{getGSISID}
\title{For a player's href, get their GSIS ID from their personal url.}
\usage{
getGSISID(value)
}
\description{
For a player's href, get their GSIS ID from their personal url.
}
