% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameandRosterFunctions.R
\name{getPageNumbers}
\alias{getPageNumbers}
\title{Get Number of Player Position Pages}
\usage{
getPageNumbers(value)
}
\description{
For each position, this function extracts the number of pages 
there are to scrape. This is a sub-function for the season_rosters function
}
