% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameandRosterFunctions.R
\name{getPlayers}
\alias{getPlayers}
\title{Scrape Player Names and Positions}
\usage{
getPlayers(position, season, type = c("REG", "POST", "PRE"))
}
\description{
This sub-function, calls buildNameAbbr and getPageNumbers to
scrape player positions by season.
}
