% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{nflteams}
\alias{nflteams}
\title{NFL Team Names and Abbreviations}
\format{This dataset contains the full team names and abbreviations
of each of the 32 teams in the NFL.  The data set has 32 rows and 2 columns}
\usage{
nflteams
}
\description{
NFL Team Names and Abbreviations
}
\keyword{datasets}
