% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlayerGameFunctions.R
\name{playergame}
\alias{playergame}
\title{Boxscore for a Single Game}
\usage{
playergame(URLString)
}
\arguments{
\item{URLString}{(character) A URL character string linking to the JSON API
data from a single game}
}
\value{
This function outputs a single dataframe containing all rushing, passing,
receiving, fumble, and defensive statistics for each player in a single  game.
Each player is assigned one line associated wih their statisitcs.
}
\description{
This function is used to neatly read all a players measurable
statistics from a given game.  Each player's statistics can be viewed on one
line.
}
\details{
This dataframe includes 36 variables including identifiers such as:
 \itemize{
 \item{"gameID", "date", "team", "playerID", "name"}
 }
Statistics are included for passing, rushing, receiving, defensive,
and fumbles.
}
\examples{
# URL Link to NFL JSON data for a random game
nfl.data <- "http://www.nfl.com/liveupdate/game-center/2013090800/2013090800_gtd.json"
PlayerGameTest <- playergame(nfl.data)
}

