% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{playerstats12}
\alias{playerstats12}
\title{NFL Team Names and Abbreviations}
\format{Outputs of the season_playergame() function for the 2012 season.  
This datset used in the vingette for nflscrapR.}
\usage{
playerstats12
}
\description{
NFL Team Names and Abbreviations
}
\keyword{datasets}
