% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDocumentation.R
\docType{data}
\name{playerstats14}
\alias{playerstats14}
\title{NFL Team Names and Abbreviations}
\format{Outputs of the season_playergame() function for the 2014 season.  
This datset used in the vingette for nflscrapR.}
\usage{
playerstats14
}
\description{
NFL Team Names and Abbreviations
}
\keyword{datasets}
