% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_games_and_urls.R
\name{scrape_game_ids}
\alias{scrape_game_ids}
\title{Scrape game ids for a given NFL season (either pre, regular, or post season)}
\usage{
scrape_game_ids(season, type = "reg")
}
\arguments{
\item{season}{Numeric 4-digit year associated with an NFL season}

\item{type}{String indicating the type of game ids to find, must either be
"pre", "reg", or "post" (default is "reg").}

\item{playoffs}{Boolean (TRUE or FALSE) if you want the playoffs game ids 
instead of the regular season (default is FALSE returning regular season games).}
}
\value{
String vector of NFL game ids from the specified season and game types.
}
\description{
This function returns all available game ids for a given season,
with an option specifying which type of game ids to return (either the 
pre, regular, or post season).
}
\examples{
# Scraping all game ids from 2017 regular season:
scrape_game_ids(2017) 
}
