% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_play_by_play.R
\name{scrape_json_play_by_play}
\alias{scrape_json_play_by_play}
\title{Scrape an individual game's JSON play-by-play data from NFL.com}
\usage{
scrape_json_play_by_play(game_id, check_url = 1)
}
\arguments{
\item{game_id}{Ten digit game id (either numeric or character) associated 
with a given NFL game.}

\item{check_url}{Indicator for whether or not to check if the game's url 
exists (default is 1, meaning yes). This is used to work with the 
\code{\link{scrape_season_play_by_play}} function that will perform its own
check prior to scraping individual games.}
}
\value{
Data frame where each individual row represents a single play in 
the game containing the following detailed information:
\itemize{
\item{play_id} - Numeric play id that when used with game_id and drive provides
the unique identifier for a single play.
\item{game_id} - Ten digit identifier for NFL game.
\item{home_team} - String abbreviation for the home team.
\item{away_team} - String abbreviation for the away team.
\item{posteam} - String abbreviation for the team with possession.
\item{posteam_type} - String indicating whether the posteam team is home or away.
\item{defteam} - String abbreviation for the team on defense.
\item{side_of_field} - String abbreviation for which team's side of the field
the team with possession is currently on.
\item{yardline_100} - Numeric distance in the number of yards from the opponent's
endzone for the posteam.
\item{game_date} - Date of the game.
\item{quarter_seconds_remaining} - Numeric seconds remaining in the quarter.
\item{half_seconds_remaining} - Numeric seconds remaining in the half.
\item{game_seconds_remaining} - Numeric seconds remaining in the game.
\item{game_half} - String indicating which half the play is in, either Half1, 
Half2, or Overtime.
\item{quarter_end} - Binary indicator for whether or not the row of the data
is marking the end of a quarter.
\item{drive} - Numeric drive number in the game.
\item{sp} - Binary indicator for whether or not a score occurred on the play.
\item{qtr} - Quarter of the game (5 is overtime).
\item{down} - The down for the given play.
\item{goal_to_go} - Binary indicator for whether or not the posteam is in a 
goal down situation.
\item{time} - Time at start of play provided in string format as minutes:seconds 
remaining in the quarter.
\item{yrdln} - String indicating the current field position for a given play.
\item{ydstogo} - Numeric yards in distance from either the first down marker or
the endzone in goal down situations.
\item{ydsnet} - Numeric value for total yards gained on the given drive.
\item{desc} - Detailed string description for the given play.
\item{play_type} - String indicating the type of play: pass (includes sacks), 
run (includes scrambles), punt, field_goal, kickoff, extra_point, 
qb_kneel, qb_spike, no_play (timeouts and penalties), and missing for rows
indicating end of play.
\item{yards_gained} - Numeric yards gained (or lost) for the given play.
\item{shotgun} - Binary indicator for whether or not the play was in shotgun
formation.
\item{no_huddle} - Binary indicator for whether or not the play was in no_huddle formation.
\item{qb_dropback} - Binary indicator for whether or not the QB dropped back on
the play (pass attempt, sack, or scrambled).
\item{qb_kneel} - Binary indicator for whether or not the QB took a knee.
\item{qb_spike} - Binary indicator for whether or not the QB spiked the ball.
\item{qb_scramble} - Binary indicator for whether or not the QB scrambled.
\item{pass_length} - String indicator for pass length: short or deep.
\item{pass_location} - String indicator for pass location: left, middle, or right.
\item{air_yards} - Numeric value for distance in yards perpendicular to the line
of scrimmage at where the targeted receiver either caught or didn't catch the ball.
\item{yards_after_catch} - Numeric value for distance in yards perpendicular to
the yard line where the receiver made the reception to where the play ended.
\item{run_location} - String indicator for location of run: left, middle, or right.
\item{run_gap} - String indicator for line gap of run: end, guard, or tackle
\item{field_goal_result} - String indicator for result of field goal attempt:
made, missed, or blocked.
\item{kick_distance} - Numeric distance in yards for kickoffs, field goals, 
and punts.
\item{extra_point_result} - String indicator for the result of the extra point
attempt: good, failed, blocked, safety (touchback in defensive endzone is 1
point apparently), or aborted.
\item{two_point_conv_result} - String indicator for result of two point conversion
attempt: success, failure, safety (touchback in defensive endzone is 1 point 
apparently), or return.
\item{home_timeouts_remaining} - Numeric timeouts remaining in the half
for the home team.
\item{away_timeouts_remaining} - Numeric timeouts remaining in the half for
the away team.
\item{timeout} - Binary indicator for whether or not a timeout was called.
\item{timeout_team} - String abbreviation for which team called the timeout.
\item{td_team} - String abbreviation for which team scored the touchdown.
\item{posteam_timeouts_remaining} - Number of timeouts remaining for the 
possession team.
\item{defteam_timeouts_remaining} - Number of timeouts remaining for the 
team on defense.
\item{total_home_score} - Score for the home team at the start of the play.
\item{total_away_score} - Score for the away team at the start of the play.
\item{posteam_score} - Score the posteam at the start of the play.
\item{defteam_score} - Score the defteam at the start of the play.
\item{score_differential} - Score differential between the posteam and defteam
at the start of the play.
\item{posteam_score_post} - Score for the posteam at the end of the play.
\item{defteam_score_post} - Score for the defteam at the end of the play.
\item{score_differential_post} - Score differential between the posteam and 
defteam at the end of the play.
\item{no_score_prob} - Predicted probability of no score occurring for the rest
of the half based on the expected points model.
\item{opp_fg_prob} - Predicted probability of the defteam scoring a FG next.
\item{opp_safety_prob} - Predicted probability of the defteam scoring a safety
next.
\item{opp_td_prob} - Predicted probability of the defteam scoring a TD next.
\item{fg_prob} - Predicted probability of the posteam scoring a FG next.
\item{safety_prob} - Predicted probability of the posteam scoring a safety next.
\item{td_prob} - Predicted probability of the posteam scoring a TD next.
\item{extra_point_prob} - Predicted probability of the posteam scoring an extra point.
\item{two_point_conversion_prob} - Predicted probability of the posteam scoring
the two point conversion.
\item{ep} - Using the scoring event probabilities, the estimated expected points
with respect to the possession team for the given play.
\item{epa} - Expected points added (EPA) by the posteam for the given play.
\item{total_home_epa} - Cumulative total EPA for the home team in the game so far.
\item{total_away_epa} - Cumulative total EPA for the away team in the game so far.
\item{total_home_rush_epa} - Cumulative total rushing EPA for the home team in
the game so far.
\item{total_away_rush_epa} - Cumulative total rushing EPA for the away team in
the game so far.
\item{total_home_pass_epa} - Cumulative total passing EPA for the home team in
the game so far.
\item{total_away_pass_epa} - Cumulative total passing EPA for the away team in
the game so far.
\item{air_epa} - EPA from the air yards alone. For completions this represents
the actual value provided through the air. For incompletions this represents
the hypothetical value that could've been added through the air if the pass
was completed.
\item{yac_epa} - EPA from the yards after catch alone. For completions this
represents the actual value provided after the catch. For incompletions this
represents the difference between the hypothetical air_epa and the play's raw
observed EPA (how much the incomplete pass cost the posteam).
\item{comp_air_epa} - EPA from the air yards alone only for completions.
\item{comp_yac_epa} - EPA from the yards after catch alone only for completions.
\item{total_home_comp_air_epa} - Cumulative total completions air EPA for the home team in
the game so far.
\item{total_away_comp_air_epa} - Cumulative total completions air EPA for the away team in
the game so far.
\item{total_home_comp_yac_epa} - Cumulative total completions yac EPA for the home team in
the game so far.
\item{total_away_comp_yac_epa} - Cumulative total completions yac EPA for the away team in
the game so far.
\item{total_home_raw_air_epa} - Cumulative total raw air EPA for the home team in
the game so far.
\item{total_away_raw_air_epa} - Cumulative total raw air EPA for the away team in
the game so far.
\item{total_home_raw_yac_epa} - Cumulative total raw yac EPA for the home team in
the game so far.
\item{total_away_raw_yac_epa} - Cumulative total raw yac EPA for the away team in
the game so far.
\item{wp} - Estimated win probabiity for the posteam given the current situation
at the start of the given play.
\item{def_wp} - Estimated win probability for the defteam.
\item{home_wp} - Estimated win probability for the home team.
\item{away_wp} - Estimated win probability for the away team.
\item{wpa} - Win probability added (WPA) for the posteam.
\item{home_wp_post} - Estimated win probability for the home team at the start of the play.
\item{away_wp_post} - Estimated win probability for the away team at the start of the play.
\item{total_home_rush_wpa} - Cumulative total rushing WPA for the home team in
the game so far.
\item{total_away_rush_wpa} - Cumulative total rushing WPA for the away team in
the game so far.
\item{total_home_pass_wpa} - Cumulative total passing WPA for the home team in
the game so far.
\item{total_away_pass_wpa} - Cumulative total passing WPA for the away team in
the game so far.
\item{air_wpa} - WPA through the air (same logic as air_epa).
\item{yac_wpa} - WPA from yards after the catch (same logic as yac_epa).
\item{comp_air_wpa} - The air_wpa for completions only.
\item{comp_yac_wpa} - The yac_wpa for completions only.
\item{total_home_comp_air_wpa} - Cumulative total completions air WPA for the home team in
the game so far.
\item{total_away_comp_air_wpa} - Cumulative total completions air WPA for the away team in
the game so far.
\item{total_home_comp_yac_wpa} - Cumulative total completions yac WPA for the home team in
the game so far.
\item{total_away_comp_yac_wpa} - Cumulative total completions yac WPA for the away team in
the game so far.
\item{total_home_raw_air_wpa} - Cumulative total raw air WPA for the home team in
the game so far.
\item{total_away_raw_air_wpa} - Cumulative total raw air WPA for the away team in
the game so far.
\item{total_home_raw_yac_wpa} - Cumulative total raw yac WPA for the home team in
the game so far.
\item{total_away_raw_yac_wpa} - Cumulative total raw yac WPA for the away team in
the game so far.
\item{punt_blocked} - Binary indicator for if the punt was blocked.
\item{first_down_rush} - Binary indicator for if a running play converted the first down.
\item{first_down_pass} - Binary indicator for if a passing play converted the first down.
\item{first_down_penalty} - Binary indicator for if a penalty converted the first down.
\item{third_down_converted} - Binary indicator for if the first down was converted on third down.
\item{third_down_failed} - Binary indicator for if the posteam failed to convert first down on third down.
\item{fourth_down_converted} - Binary indicator for if the first down was converted on fourth down.
\item{fourth_down_failed} - Binary indicator for if the posteam failed to convert first down on fourth down.
\item{incomplete_pass} - Binary indicator for if the pass was incomplete.
\item{interception} - Binary indicator for if the pass was intercepted.
\item{touchback} - Binary indicator for if a touchback occurred on the play.
\item{punt_inside_twenty} - Binary indicator for if the punt ended inside the twenty yard line.
\item{punt_in_endzone} - Binary indicator for if the punt was in the endzone.
\item{punt_out_of_bounds} - Binary indicator for if the punt went out of bounds.
\item{punt_downed} - Binary indicator for if the punt was downed.
\item{punt_fair_catch} - Binary indicator for if the punt was caught with a fair catch.
\item{kickoff_inside_twenty} - Binary indicator for if the kickoff ended inside the twenty yard line.
\item{kickoff_in_endzone} - Binary indicator for if the kickoff was in the endzone.
\item{kickoff_out_of_bounds} - Binary indicator for if the kickoff went out of bounds.
\item{kickoff_downed} - Binary indicator for if the kickoff was downed.
\item{kickoff_fair_catch} - Binary indicator for if the kickoff was caught with a fair catch.
\item{fumble_forced} - Binary indicator for if the fumble was forced.
\item{fumble_not_forced} - Binary indicator for if the fumble was not forced.
\item{fumble_out_of_bounds} - Binary indicator for if the fumble went out of bounds.
\item{solo_tackle} - Binary indicator if the play had a solo tackle (could be multiple
due to fumbles).
\item{safety} - Binary indicator for whether or not a safety occurred.
\item{penalty} - Binary indicator for whether or not a penalty occurred.
\item{tackled_for_loss} - Binary indicator for whether or not a tackle for loss occurred.
\item{fumble_lost} - Binary indicator for if the fumble was lost.
\item{own_kickoff_recovery} - Binary indicator for if the kicking team recovered the
kickoff.
\item{own_kickoff_recovery_td} - Binary indicator for if the kicking team recovered
the kickoff and scored a TD.
\item{qb_hit} - Binary indicator if the QB was hit on the play.
\item{rush_attempt} - Binary indicator for if the play was a run.
\item{pass_attempt} - Binary indicator for if the play was a pass attempt (includes sacks).
\item{sack} - Binary indicator for if the play ended in a sack.
\item{touchdown} - Binary indicator for if the play resulted in a TD.
\item{pass_touchdown} - Binary indicator for if the play resulted in a passing TD.
\item{rush_touchdown} - Binary indicator for if the play resulted in a rushing TD.
\item{return_touchdown} - Binary indicator for if the play resulted in a return TD.
\item{extra_point_attempt} - Binary indicator for extra point attempt.
\item{two_point_attempt} - Binary indicator for two point conversion attempt.
\item{field_goal_attempt} - Binary indicator for field goal attempt.
\item{kickoff_attempt} - Binary indicator for kickoff.
\item{punt_attempt} - Binary indicator for punts.
\item{fumble} - Binary indicator for if a fumble occurred.
\item{complete_pass} - Binary indicator for if the pass was completed.
\item{assist_tackle} - Binary indicator for if an assist tackle occurred.
\item{lateral_reception} - Binary indicator for if a lateral occurred on the
reception.
\item{lateral_rush} - Binary indicator for if a lateral occurred on a run.
\item{lateral_return} - Binary indicator for if a lateral occurred on a return.
\item{lateral_recovery} - Binary indicator for if a lateral occurred on a 
fumble recovery.
\item{passer_player_id} - Unique identifier for the player that attempted the pass.
\item{passer_player_name} - String name for the player that attempted the pass.
\item{receiver_player_id} - Unique identifier for the receiver that was targeted on the pass.
\item{receiver_player_name} - String name for the targeted receiver.
\item{rusher_player_id} - Unique identifier for the player that attempted the run.
\item{rusher_player_name} - String name for the player that attempted the run.
\item{lateral_receiver_player_id} - Unique identifier for the player that received the lateral on a reception.
\item{lateral_receiver_player_name} - String name for the player that received the lateral on a reception.
\item{lateral_rusher_player_id} - Unique identifier for the player that received the lateral on a run.
\item{lateral_rusher_player_name} - String name for the player that received the lateral on a run.
\item{lateral_sack_player_id} - Unique identifier for the player that received the lateral on a sack.
\item{lateral_sack_player_name} - String name for the player that received the lateral on a sack.
\item{interception_player_id} - Unique identifier for the player that intercepted the pass.
\item{interception_player_name} - String name for the player that intercepted the pass.
\item{lateral_interception_player_id} - Unique indentifier for the player that received the lateral on an interception.
\item{lateral_interception_player_name} - String name for the player that received the lateral on an interception.
\item{punt_returner_player_id} - Unique identifier for the punt returner.
\item{punt_returner_player_name} - String name for the punt returner.
\item{lateral_punt_returner_player_id} - Unique identifier for the player that received the lateral on a punt return.
\item{lateral_punt_returner_player_name} - String name for the player that received the lateral on a punt return.
\item{kickoff_returner_player_name} - String name for the kickoff returner.
\item{kickoff_returner_player_id} - Unique identifier for the kickoff returner.
\item{lateral_kickoff_returner_player_id} - Unique identifier for the player that received the lateral on a kickoff return.
\item{lateral_kickoff_returner_player_name} - String name for the player that received the lateral on a kickoff return.
\item{punter_player_id} - Unique identifier for the punter. 
\item{punter_player_name} - String name for the punter.
\item{kicker_player_name} - String name for the kicker on FG or kickoff.
\item{kicker_player_id} - Unique identifier for the kicker on FG or kickoff.
\item{own_kickoff_recovery_player_id} - Unique identifier for the player that recovered their own kickoff.
\item{own_kickoff_recovery_player_name} - String name for the player that recovered their own kickoff.
\item{blocked_player_id} - Unique identifier for the player that blocked the punt or FG.
\item{blocked_player_name} - String name for the player that blocked the punt or FG.
\item{tackle_for_loss_1_player_id} - Unique identifier for one of the potential players with the tackle for loss.
\item{tackle_for_loss_1_player_name} - String name for one of the potential players with the tackle for loss.
\item{tackle_for_loss_2_player_id} - Unique identifier for one of the potential players with the tackle for loss.
\item{tackle_for_loss_2_player_name} - String name for one of the potential players with the tackle for loss.
\item{qb_hit_1_player_id} - Unique identifier for one of the potential players that hit the QB.
\item{qb_hit_1_player_name} - String name for one of the potential players that hit the QB.
\item{qb_hit_2_player_id} - Unique identifier for one of the potential players that hit the QB.
\item{qb_hit_2_player_name} - String name for one of the potential players that hit the QB.
\item{forced_fumble_player_1_team} - Team of one of the players with a forced fumble.
\item{forced_fumble_player_1_player_id} - Unique identifier of one of the players with a forced fumble.
\item{forced_fumble_player_1_player_name} - String name of one of the players with a forced fumble.
\item{forced_fumble_player_2_team} - Team of one of the players with a forced fumble.
\item{forced_fumble_player_2_player_id} - Unique identifier of one of the players with a forced fumble.
\item{forced_fumble_player_2_player_name} - String name of one of the players with a forced fumble.
\item{solo_tackle_1_team} - Team of one of the players with a solo tackle.
\item{solo_tackle_2_team} - Team of one of the players with a solo tackle.
\item{solo_tackle_1_player_id} - Unique identifier of one of the players with a solo tackle.
\item{solo_tackle_2_player_id} - Unique identifier of one of the players with a solo tackle.
\item{solo_tackle_1_player_name} - String name of one of the players with a solo tackle.
\item{solo_tackle_2_player_name} - String name of one of the players with a solo tackle.
\item{assist_tackle_1_player_id} - Unique identifier of one of the players with a tackle assist.
\item{assist_tackle_1_player_name} - String name of one of the players with a tackle assist.
\item{assist_tackle_1_team} - Team of one of the players with a tackle assist.
\item{assist_tackle_2_player_id} - Unique identifier of one of the players with a tackle assist.
\item{assist_tackle_2_player_name} - String name of one of the players with a tackle assist.
\item{assist_tackle_2_team} - Team of one of the players with a tackle assist.
\item{assist_tackle_3_player_id} - Unique identifier of one of the players with a tackle assist.
\item{assist_tackle_3_player_name} - String name of one of the players with a tackle assist.
\item{assist_tackle_3_team} - Team of one of the players with a tackle assist.
\item{assist_tackle_4_player_id} - Unique identifier of one of the players with a tackle assist.
\item{assist_tackle_4_player_name} - String name of one of the players with a tackle assist.
\item{assist_tackle_4_team} - Team of one of the players with a tackle assist.
\item{pass_defense_1_player_id} - Unique identifier of one of the players with a pass defense.
\item{pass_defense_1_player_name} - String name of one of the players with a pass defense.
\item{pass_defense_2_player_id} - Unique identifier of one of the players with a pass defense.
\item{pass_defense_2_player_name} - String name of one of the players with a pass defense.
\item{fumbled_1_team} - Team of one of the players with a fumble.
\item{fumbled_1_player_id} - Unique identifier of one of the players with a fumble.
\item{fumbled_1_player_name} - String name of one of the players with a fumble.
\item{fumbled_2_player_id} - Unique identifier of one of the players with a fumble.
\item{fumbled_2_player_name} - String name of one of the players with a fumble.
\item{fumbled_2_team} - Team of one of the players with a fumble.
\item{fumble_recovery_1_team} - Team of one of the players with a fumble recovery.
\item{fumble_recovery_1_yards} - Yards gained by one of the players with a fumble recovery.
\item{fumble_recovery_1_player_id} - Unique identifier of one of the players with a fumble recovery.
\item{fumble_recovery_1_player_name} - String name of one of the players with a fumble recovery.
\item{fumble_recovery_2_team} - Team of one of the players with a fumble recovery.
\item{fumble_recovery_2_yards} - Yards gained by one of the players with a fumble recovery.
\item{fumble_recovery_2_player_id} - Unique identifier of one of the players with a fumble recovery.
\item{fumble_recovery_2_player_name} - String name of one of the players with a fumble recovery.
\item{return_team} - String abbreviation of the return team.
\item{return_yards} - Yards gained by the return team.
\item{penalty_team} - String abbreviation of the team with the penalty.
\item{penalty_player_id} - Unique identifier for the player with the penalty.
\item{penalty_player_name} - String name for the player with the penalty.
\item{penalty_yards} - Yards gained (or lost) by the posteam from the penalty.
\item{replay_or_challenge} - Binary indicator for whether or not a replay or
challenge.
\item{replay_or_challenge_result} - String indicating the result of the 
replay or challenge.
\item{penalty_type} - String indicating the penalty type.
\item{defensive_two_point_attempt} - Binary indicator whether or not the defense
was able to have an attempt on a two point conversion, this results following
a turnover.
\item{defensive_two_point_conv} - Binary indicator whether or not the defense
successfully scored on the two point conversion.
\item{defensive_extra_point_attempt} - Binary indicator whether or not the 
defense was able to have an attempt on an extra point attempt, this results
following a blocked attempt that the defense recovers the ball.
\item{defensive_extra_point_conv} - Binary indicator whether or not the
defense successfully scored on an extra point attempt. 
}
}
\description{
This function returns the play-by-play data available from NFL.com's JSON feed 
(for games starting in 2009).
}
\examples{
# Scrape the play-by-play data for the 2017 Super Bowl by first getting all
# of the post-season game ids then use the required info to scrape the 
# play-by-play data (knowing that it's the last game):
playoff_game_ids_17 <- scrape_game_ids(2017, type = "post")
sb_17_id <- playoff_game_ids_17$game_id[nrow(playoff_game_ids_17)]
sb_17_pbp <- scrape_json_play_by_play(game_id = sb_17_id)
}
