% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameandRosterFunctions.R
\name{season_games}
\alias{season_games}
\title{Game Information for All Games in a Season}
\usage{
season_games(Season)
}
\arguments{
\item{Season}{(numeric) A 4-digit year associated with a given NFL season}
}
\value{
A dataframe with the gameID, the game date, 
the home team abbreviation, and away team abbreviation
}
\description{
This function intakes a year associated with a given season
and outputs all the game matchups for all 17 weeks of the regular season
}
\details{
Reference the stored dataframe nflteams to match team abbreviations
with the full team names
}
\examples{
# All games in 2015 Season
season_games(2015) # Will output a dataframe
}

