% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlayerGameFunctions.R
\name{season_playergame}
\alias{season_playergame}
\title{Boxscore for Each Game in the Season - One line per player per game}
\usage{
season_playergame(Season)
}
\arguments{
\item{Season}{(numeric) A 4-digit year associated with a given season}
}
\value{
A single line for each player in each game they played in.  The 
output is the same as the playergame function but is run for every game in
the specified season
}
\description{
This function outputs a single dataframe containing all rushing, 
passing, receiving, fumble, and defensive statistics for each player in 
each game.  Each player is assigned one line associated wih their statisitcs 
per each game they record a measured statistic
}
\examples{
# Player-Game function over the entire season in 2010
playerstats.2010 <- season_playergame(2010)
head(playerstats.2010)

# Plot a graph of different play types
ggplot(playerstats.2010, aes(x = PlayType)) + geom_bar()
}

