% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameandRosterFunctions.R
\name{season_rosters}
\alias{season_rosters}
\title{Season Rosters for Teams}
\usage{
season_rosters(Season, TeamInt)
}
\arguments{
\item{Season:}{A 4-digit year associated with a given NFL season}

\item{TeamInt:}{A string containing the abbreviations for an NFL Team}
}
\value{
A dataframe with columns associated with season/year, team, playerID,
players who played and recorded some measurable statistic, and the 
last column specifyng the number of games they played in.
}
\description{
This function intakes a year and a team abbreviation and outputs
a dataframe with each player who has played for the specified team and 
recorded a measurable statistic
}
\details{
To find team associated abbrevations use the nflteams dataframe 
stored in this package!
}
\examples{
# Roster for Baltimore Ravens in 2013
season_rosters(2013, TeamInt = "BAL") 
}

