% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlayByPlayBoxScore.R
\name{simple_boxscore}
\alias{simple_boxscore}
\title{Drive Summary and Results}
\usage{
simple_boxscore(URLString, home = TRUE)
}
\arguments{
\item{URLString}{(character) is a character string of the URL of where to
pull the data from}

\item{home}{(boolean): home = TRUE will pull home stats,
home = FALSE pulls away stats}
}
\value{
A list of statistics including passing, rushing, receiving, defense,
 kicking, kick return, and punt return statistics for the specified game
}
\description{
This function pulls data from an NFL url and contructs it into a formatted
boxscore.
}
\examples{
# Parsed drive Summarize of final game in 2015 NFL Season
nfl2015.finalregseasongame <- "http://www.nfl.com/liveupdate/game-center/2016010310/2016010310_gtd.json"
simple_BoxScore(nfl2015.finalregseasongame, home = TRUE)
}

