% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpectedPointAndWinProbability.R
\name{win_probability}
\alias{win_probability}
\title{Win probability function to add win probability columns for the offense and 
defense for each play in the game}
\usage{
win_probability(dataset)
}
\arguments{
\item{dataset}{(data.frame object) A data.frame as exported from the 
game_play_by_play function}
}
\value{
The input dataframe with the addition of win probability columns for 
both the offense and defense
}
\description{
This function takes in the output of the game level play by play
function and returns the same dataframe with two new columns representing the
win probability for the offensive and defensive teams.  The model used to 
calculate win probability can be further examined by loading it into your
workspace from this package.
}

