% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlayByPlayBoxScore.R
\name{game_play_by_play}
\alias{game_play_by_play}
\title{Parsed Descriptive Play-by-Play Dataset for a Single Game}
\usage{
game_play_by_play(GameID)
}
\arguments{
\item{GameID}{(character or numeric) A 10 digit game ID associated with a 
given NFL game.}
}
\value{
A dataframe with 62 columns specifying various statistics and 
outcomes associated with each play of the specified NFL game.
}
\description{
This function intakes the JSON play-by-play data of a single
 game and parses the play description column into individual variables 
 allowing the user to segment the game in a variety of different ways for 
 model building and analysis.
}
\details{
Through list manipulation using the do.call and rbind functions
 a 10 column dataframe with basic information populates directly from the NFL 
 JSON API.  These columns include the following:
\itemize{
 \item{"Drive"} - Drive number
 \item{"sp"} - Whether the play resulted in a score (any kind of score)
 \item{"qrt"} - Quarter of Game
 \item{"down"} - Down of the given play
 \item{"time"} - Time at start of play
 \item{"yrdln"} - Between 0 and 50
 \item{"ydstogo"} - For a first down
 \item{"ydsnet"} - Total yards gained on a given drive
 \item{"posteam"} - The offensive team
 \item{"desc"} - A detailed description of what occured during the play
}

Through string manipulation and parsing of the description column using the 
base R and stringR, 51 columns were added to the original dataframe allowing 
the user to have a detailed breakdown of the events of each play.  
The added variables are specified below:
\itemize{
 \item{"Date"} - Date of game
 \item{"GameID"} - The ID of the specified game
 \item{"TimeSecs"} - Time remaining in game in seconds
 \item{"PlayTimeDiff"} - The time difference between plays in seconds
 \item{"DefensiveTeam"} - The defensive team on the play (for punts the 
 receiving team is on defense, for kickoffs the receiving team is on offense)
 \item{"TimeUnder"} - 
 \item{"SideofField"} - The side of the field that the line of scrimmage 
 is on
 \item{yrdline100} - Distance to opponents enzone, ranges from 1-99.
 situation
 \item{GoalToGo} - Binary variable indicting if the play is in a goal-to-go
 situation
 \item{"FirstDown"} - Binary: 0 if the play did not result in a first down 
 and 1 if it did
 \item{"PlayAttempted"} - A variabled used to count the number of plays in a 
 game (should always be equal to 1)
 \item{"Yards.Gained"} - Amount of yards gained on the play
 \item{"Touchdown"} - Binary: 1 if the play resulted in a TD else 0
 \item{"ExPointResult"} - Result of the extra-point: Made, Missed, Blocked
 \item{"TwoPointConv"} - Result of two-point conversion: Success of Failure
 \item{"DefTwoPoint"} - Result of defesnive two-point conversion: Success of Failure
 \item{"Safety"} - Binary: 1 if safety was recorded else 0
 \item{"Onsidekick"} - Binary: 1 if the Kickoff was an onside kick
 \item{"PuntResult} - The resulting action of a punt.  Either a clean punt or a blocked punt
 \item{"PlayType"} - The type of play that occured. Potential values are:
       \itemize{
                 \item{Kickoff, Punt, Onside Kick}
                 \item{Passs, Run}
                 \item{Sack}
                 \item{Field Goal, Extra Point}
                 \item{Quarter End, Two Minute Warning, End of Game}
                 \item{No Play, QB Kneel, Spike, Timeout}
         }  
 \item{"Passer"} - The passer on the play if it was a pass play
 \item{"PassAttempt"} - Binary variable indicating whether a pass was attempted
 or not
 \item{"PassOutcome"} - Pass Result: Complete or Incomplete    
 \item{"PassLength"} - Categorical variable indicating the length of the pass:
 Short or Deep
 \item{"PassLocation"} - Categorical variable: left, middle, right
 \item{"InterceptionThrown"} - Binary variable indicating whether an 
 interception was thrown
 \item{"Interceptor"} - The player who intercepted the ball
 \item{"Rusher"} - The runner on the play if it was a running play
 \item{"RushAttempt"} - Binary variable indicating whether or not a run was 
 attempted.
 \item{"RunLocation"} - The location of the run - left, middle, right
 \item{"RunGap"} - The gap that the running back ran through
 \item{"Receiver"} - The player who recorded the reception on a complete pass
 \item{"Reception"} - Binary Variable indicating a reception on a completed 
 pass: 1 if a reception was recorded else 0
 \item{"ReturnResult"} - Result of a punt, kickoff, interception, or 
 fumble return
 \item{"Returner"} - The punt or kickoff returner
 \item{"BlockingPlayer"} - The player who blocked the extra point, 
 field goal, or punt
 \item{"Tackler1"} - The primary tackler on the play
 \item{"Tackler2"} - The secondary tackler on the play
 \item{"FieldGoalResult"} - Outcome of a fieldgoal: made, missed, blocked
 \item{"FieldGoalDistance"} - Field goal length in yards
 \item{"Fumble"} - Binary variable indicating whether a fumble occured or not:
 1 if a fumble occured else no 
 \item{"RecFumbTeam"} - Team that recovered the fumble
 \item{"RecFumbPlayer"} - Player that recovered the fumble
 \item{"Sack"} - Binary variable indicating whether a sack was recorded: 1 if
 a sack was recorded else 0
 \item{"Challenge.Replay"} - Binary variable indicating whether or not the 
 play was reviewed by the replay offical on challenges or replay reviews
 \item{"ChalReplayResult"} - Result of the replay review: Upheld or Overturned
 \item{"Accepted.Penalty"} - Binary variable indicating whether a penalty was 
 accpeted on the play
 \item{"PenalizedTeam"} - The team who was penalized on the play
 \item{"PenaltyType"} - Type of penalty on the play. Values include:
       \itemize{    
                 \item{Unnecessary Roughness, Roughing the Passer}
                 \item{Illegal Formation, Defensive Offside}
                 \item{Delay of Game, False Start, Illegal Shift}
                 \item{Illegal Block Above the Waist, Personal Foul}
                 \item{Unnecessary Roughness, Illegal Blindside Bloc}
                 \item{Defensive Pass Interference, Offensive Pass Interference}
                 \item{Fair Catch Interferenc, Unsportsmanlike Conduct}
                 \item{Running Into the Kicker, Illegal Kick}
                 \item{Illegal Contact, Defensive Holding}
                 \item{Illegal Motion, Low Block}
                 \item{Illegal Substitution, Neutral Zone Infraction}
                 \item{Ineligible Downfield Pass, Roughing the Passer}
                 \item{Illegal Use of Hands, Defensive Delay of Game}
                 \item{Defensive 12 On-field, Offensive Offside}
                 \item{Tripping, Taunting, Chop Block}
                 \item{Interference with Opportunity to Catch, Illegal Touch Pass}
                 \item{Illegal Touch Kick, Offside on Free Kick}
                 \item{Intentional Grounding, Horse Collar}
                 \item{Illegal Forward Pass, Player Out of Bounds on Punt}
                 \item{Clipping, Roughing the Kicker, Ineligible Downfield Kick}
                 \item{Offensive 12 On-field, Disqualification}
      } 
 \item{"PenalizedPlayer"} - The penalized player
 \item{"Penalty.Yards"} - The number of yards that the penalty resulted in
 \item{"PosTeamScore"} - The score of the possession team (offensive team)
 \item{"DefTeamScore"} - The score of the defensive team
 \item{"ScoreDiff"} - The difference in score between the offensive and 
 defensive teams (offensive.score - def.score)         
 \item{"AbsScoreDiff"} - The absolute score difference on the given play
 
 }
}
\examples{
# Parsed play-by-play of the final game in the 2015 NFL season 

# Save the gameID into a variable 
nfl2015.finalregseasongame.gameID <- "2016010310"

# Input the variable into the function to output the desired dataframe
finalgame2015.pbp <- game_play_by_play(nfl2015.finalregseasongame.gameID) 

# Subset the dataframe based on passing plays
subset(finalgame2015.pbp, PlayType == "Pass")
}

