% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlayByPlayBoxScore.R
\name{game_play_by_play}
\alias{game_play_by_play}
\title{Parsed Descriptive Play-by-Play Dataset for a Single Game}
\usage{
game_play_by_play(GameID)
}
\arguments{
\item{GameID}{(character or numeric) A 10 digit game ID associated with a 
given NFL game.}
}
\value{
A dataframe with 99 columns specifying various statistics and 
outcomes associated with each play of the specified NFL game.
}
\description{
This function intakes the JSON play-by-play data of a single
 game and parses the play description column into individual variables 
 allowing the user to segment the game in a variety of different ways for 
 model building and analysis. WARNING: This function is deprecated and will be removed from the package
 after the 2018-19 season.
}
\details{
Through list manipulation using the do.call and rbind functions
 a 13 column dataframe with basic information populates directly from the NFL 
 JSON API.  These columns include the following:
\itemize{
 \item{"Drive"} - Drive number
 \item{"sp"} - Whether the play resulted in a score (any kind of score)
 \item{"qtr"} - Quarter of Game
 \item{"down"} - Down of the given play
 \item{"time"} - Time at start of play
 \item{"yrdln"} - Between 0 and 50
 \item{"ydstogo"} - Yards to go for a first down
 \item{"ydsnet"} - Total yards gained on a given drive
 \item{"posteam"} - The team on offense
 \item{"AirYards"} - Number of yards the ball was thrown in the air for both
 complete and incomplete pass attempts (negative means behind line of scrimmage)
 \item{"YardsAfterCatch"} - Number of yards receiver gained after catch
 \item{"QBHit"} -  Binary: 1 if the QB was knocked to the ground else 0
 \item{"desc"} - A detailed description of what occured during the play
}

Through string manipulation and parsing of the description column using  
base R and stringR, columns were added to the original dataframe allowing 
the user to have a detailed breakdown of the events of each play. Also
provided are calculations for the expected points and win probability
for each play using models built entirely on nflscrapR data. 
The added variables are specified below:
\itemize{
 \item{"Date"} - Date of game
 \item{"GameID"} - The ID of the specified game
 \item{"play_id"} - Play id within a game
 \item{"TimeSecs"} - Time remaining in game in seconds
 \item{"PlayTimeDiff"} - The time difference between plays in seconds
 \item{"DefensiveTeam"} - The defensive team on the play (for punts the 
 receiving team is on defense, for kickoffs the receiving team is on offense)
 \item{"TimeUnder"} - Minutes remaining in half
 \item{"SideofField"} - The side of the field that the line of scrimmage 
 is on
 \item{yrdline100} - Distance to opponents endzone, ranges from 1-99
 situation
 \item{GoalToGo} - Binary: 1 if the play is in a goal down situation else 0
 \item{"FirstDown"} - Binary: 1if the play resulted in a first down conversion
 else 0
 \item{"PlayAttempted"} - A variable used to count the number of plays in a 
 game (should always be equal to 1)
 \item{"Yards.Gained"} - Amount of yards gained on the play
 \item{"Touchdown"} - Binary: 1 if the play resulted in a TD else 0
 \item{"ExPointResult"} - Result of the extra-point: Made, Missed, Blocked, 
 Aborted
 \item{"TwoPointConv"} - Result of two-point conversion: Success or Failure
 \item{"DefTwoPoint"} - Result of defensive two-point conversion: Success or Failure
 \item{"Safety"} - Binary: 1 if safety was recorded else 0
 \item{"Onsidekick"} - Binary: 1 if the kickoff was an onside kick
 \item{"PuntResult} - Result of punt: Clean or Blocked
 \item{"PlayType"} - The type of play that occured, potential values are:
       \itemize{
                 \item{Kickoff, Punt}
                 \item{Pass, Sack, Run}
                 \item{Field Goal, Extra Point}
                 \item{Quarter End, Two Minute Warning, Half End, End of Game}
                 \item{No Play, QB Kneel, Spike, Timeout}
         }  
 \item{"Passer"} - The passer on the play if it was a pass play
 \item{"Passer_ID"} - NFL GSIS player ID for the passer
 \item{"PassAttempt"} - Binary: 1 if a pass was attempted else 0
 \item{"PassOutcome"} - Pass Result: Complete or Incomplete Pass  
 \item{"PassLength"} - Categorical variable indicating the length of the pass:
 Short or Deep
 \item{"PassLocation"} - Location of the pass: left, middle, right
 \item{"InterceptionThrown"} - Binary: 1 if an interception else 0
 \item{"Interceptor"} - The player who intercepted the ball
 \item{"Rusher"} - The runner on the play if it was a running play
 \item{"Rusher_ID"} - NFL GSIS player ID for the rusher
 \item{"RushAttempt"} - Binary: 1 if a run was attempted else 0
 \item{"RunLocation"} - Location of the run: left, middle, right
 \item{"RunGap"} - Gap of the run: guard, tackle, end 
 \item{"Receiver"} - The targeted receiver of a play
 \item{"Receiver_ID"} - NFL GSIS player ID for the receiver
 \item{"Reception"} - Binary: 1 if a reception was recorded else 0
 \item{"ReturnResult"} - Result of a punt, kickoff, interception, or 
 fumble return: Fair Catch, Touchback, Touchdown
 \item{"Returner"} - The punt or kickoff returner
 \item{"BlockingPlayer"} - The player who blocked the extra point, 
 field goal, or punt
 \item{"Tackler1"} - The primary tackler on the play
 \item{"Tackler2"} - The secondary tackler on the play
 \item{"FieldGoalResult"} - Outcome of a fieldgoal: No Good, Good, Blocked
 \item{"FieldGoalDistance"} - Field goal length in yards
 \item{"Fumble"} - Binary: 1 if a fumble occured else no 
 \item{"RecFumbTeam"} - Team that recovered the fumble
 \item{"RecFumbPlayer"} - Player that recovered the fumble
 \item{"Sack"} - Binary: 1 if a sack was recorded else 0
 \item{"Challenge.Replay"} - Binary: 1 if play was reviewed by the replay official
 else 0
 \item{"ChalReplayResult"} - Result of the replay review: Upheld or Reversed
 \item{"Accepted.Penalty"} - Binary: 1 if penalty was accepted else 0
 \item{"PenalizedTeam"} - The penalized team on the play
 \item{"PenaltyType"} - Type of penalty on the play, alues include:
       \itemize{    
                 \item{Unnecessary Roughness, Roughing the Passer}
                 \item{Illegal Formation, Defensive Offside}
                 \item{Delay of Game, False Start, Illegal Shift}
                 \item{Illegal Block Above the Waist, Personal Foul}
                 \item{Unnecessary Roughness, Illegal Blindside Block}
                 \item{Defensive Pass Interference, Offensive Pass Interference}
                 \item{Fair Catch Interference, Unsportsmanlike Conduct}
                 \item{Running Into the Kicker, Illegal Kick}
                 \item{Illegal Contact, Defensive Holding}
                 \item{Illegal Motion, Low Block}
                 \item{Illegal Substitution, Neutral Zone Infraction}
                 \item{Ineligible Downfield Pass, Roughing the Passer}
                 \item{Illegal Use of Hands, Defensive Delay of Game}
                 \item{Defensive 12 On-field, Offensive Offside}
                 \item{Tripping, Taunting, Chop Block}
                 \item{Interference with Opportunity to Catch, Illegal Touch Pass}
                 \item{Illegal Touch Kick, Offside on Free Kick}
                 \item{Intentional Grounding, Horse Collar}
                 \item{Illegal Forward Pass, Player Out of Bounds on Punt}
                 \item{Clipping, Roughing the Kicker, Ineligible Downfield Kick}
                 \item{Offensive 12 On-field, Disqualification}
      } 
 \item{"PenalizedPlayer"} - The penalized player
 \item{"Penalty.Yards"} - The number of yards that the penalty resulted in
 \item{"PosTeamScore"} - The score of the possession team (offensive team)
 \item{"DefTeamScore"} - The score of the defensive team
 \item{"ScoreDiff"} - The difference in score between the offensive and 
 defensive teams (offensive.score - def.score)
 \item{"AbsScoreDiff"} - Absolute value of the score differential
 \item{"HomeTeam"} - The home team
 \item{"AwayTeam"} - The away team
 \item{"Timeout_Indicator"} - Binary: 1 if a timeout was charge else 0
 \item{"Timeout_Team"} - Team charged with penalty (None if no timeout)
 \item{"posteam_timeouts_pre"} - Number of timeouts remaining for possession
 team at the start of the play
 \item{"HomeTimeouts_Remaining_Pre"} - Number of timeouts remaining for home
 team at the start of the play
 \item{"AwayTimeouts_Remaining_Pre"} - Number of timeouts remaining for away
 team at the start of the play
 \item{"HomeTimeouts_Remaining_Post"} - Number of timeouts remaining for home
 team at the end of the play (handles loss of timeout from lost challenge)
 \item{"AwayTimeouts_Remaining_Post"} - Number of timeouts remaining for away
 team at the end of the play (handles loss of timeout from lost challenge)
 \item{"No_Score_Prob"} - Probability of no score occurring within the half
 \item{"Opp_Field_Goal_Prob"} - Probability of the defensive team scoring a
 field goal next
 \item{"Opp_Safety_Prob"} - Probability of the defensive team scoring a 
 safety next
 \item{"Opp_Touchdown_Prob"} - Probability of the defensive team scoring a 
 touchdown next
 \item{"Field_Goal_Prob"} - Probability of the possession team scoring a 
 field goal next
 \item{"Safety_Prob"} - Probability of the possession team scoring a safety
 next
 \item{"Touchdown_Prob"} - Probability of the possession team scoring a 
 touchdown next
 \item{"ExPoint_Prob"} - Probability of the possession team making the PAT
 \item{"TwoPoint_Prob"} - Probability of the possession team converting 
 the two-point conversion
 \item{"ExpPts"} - The expected points for the possession team at the
 start of the play
 \item{"EPA"} - Expected points added with respect to the possession
 team considering the result of the play
 \item{"airEPA"} - Expected points added from air yards
 \item{"yacEPA"} - Expected points added from yards after catch
 \item{"Home_WP_pre"} - The win probability for the home team at the start
 of the play
 \item{"Away_WP_pre"} - The win probability for the away team at the start
 of the play
 \item{"Home_WP_post"} - The win probability for the home team at the
 end of the play
 \item{"Away_WP_post"} - The win probability for the away team at the
 end of the play
 \item{"Win_Prob"} - The win probability added for team with possession
 \item{"WPA"} - The win probability added with respect to the
 possession team
 \item{"airWPA"} - Win probability added from air yards
 \item{"yacWPA"} - Win probability added from yards after catch
 
 }
}
\examples{
# Parsed play-by-play of the final game in the 2015 NFL season 

# Save the gameID into a variable 
nfl2015.finalregseasongame.gameID <- "2016010310"

# Input the variable into the function to output the desired dataframe
finalgame2015.pbp <- game_play_by_play(nfl2015.finalregseasongame.gameID) 

# Subset the dataframe based on passing plays
subset(finalgame2015.pbp, PlayType == "Pass")
}
