% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScrapingFunctionsGameURLs.R
\name{proper_jsonurl_formatting}
\alias{proper_jsonurl_formatting}
\title{Formatting URL for location of NFL Game JSON Data}
\usage{
proper_jsonurl_formatting(GameID)
}
\arguments{
\item{GameID}{(character or numeric) A 10 digit game ID associated with a 
given NFL game.}
}
\value{
The url where the game JSON data for the given game can be found.
}
\description{
This function pastes together the proper formatting of 
the nfl play by play data JSON URL such that it can be used in our 
play-by-play functions.  This function calls the extracting_gameids function
}
\examples{
# Save the gameID into a variable 
nfl2015.finalregseasongame.gameID <- "2016010310"

# Putting all game IDs in to proper URL format
proper_jsonurl_formatting(nfl2015.finalregseasongame.gameID) 
}
