% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameandRosterFunctions.R
\name{season_rosters}
\alias{season_rosters}
\title{Season Rosters for Teams}
\usage{
season_rosters(season, teams, positions = c("QUARTERBACK", "RUNNING_BACK",
  "WIDE_RECEIVER", "TIGHT_END"))
}
\arguments{
\item{season:}{A 4-digit year associated with a given NFL season}

\item{teams:}{A string vector containing the abbreviations for NFL Team(s)}

\item{Positions:}{A string vector containing the abbreviations for NFL position(s).
Can be of the following:
\itemize{
\item{"QUARTERBACK"} (in the default list)
\item{"RUNNING_BACK"} (in the default list)
\item{"WIDE_RECEIVER"} (in the default list)
\item{"TIGHT_END"} (in the default list)
\item{"DEFENSIVE_LINEMAN"} 
\item{"LINEBACKER"}
\item{"DEFENSIVE_BACK"}
\item{"KICKOFF_KICKER"}
\item{"KICK_RETURNER"}
\item{"PUNTER"}
\item{"PUNT_RETURNER"}
\item{"FIELD_GOAL_KICKER"}
}}
}
\value{
A dataframe with columns associated with season/year, full player name,
team initial, position, and formated player name.
players who played and recorded some measurable statistic, and the 
last column specifyng the number of games they played in.
}
\description{
This function intakes a year and a team abbreviation and outputs
a dataframe with each player who has played for the specified team and 
recorded a measurable statistic
}
\details{
To find team associated abbrevations use the nflteams dataframe 
stored in this package!
}
\examples{
# Roster for Baltimore Ravens in 2013
season_rosters(2013, TeamInt = "BAL") 
}
