% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlayByPlayBoxScore.R
\name{simple_boxscore}
\alias{simple_boxscore}
\title{Simple Game Boxscore}
\usage{
simple_boxscore(GameID, home = TRUE)
}
\arguments{
\item{GameID}{(character or numeric) A 10 digit game ID associated with a 
given NFL game.}

\item{home}{(boolean): home = TRUE will pull home stats, 
home = FALSE pulls away stats}
}
\value{
A list of player statistics including passing, rushing, receiving, 
defense, kicking, kick return, and punt return statistics for the specified 
game.
}
\description{
This function pulls data from an NFL url and contructs it into a formatted 
boxscore.
}
\examples{
# Parsed drive summaries of final game in 2015 NFL season
nfl2015.finalregseasongame.gameID <- "2016010310"
simple_boxscore(nfl2015.finalregseasongame.gameID, home = TRUE) 
}
