% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpectedPointAndWinProbability.R
\name{win_probability}
\alias{win_probability}
\title{Win probability function to add win probability columns for the offense and 
defense for each play in the game}
\usage{
win_probability(dataset)
}
\arguments{
\item{dataset}{(data.frame object) A data.frame as exported from the 
game_play_by_play function}
}
\value{
The input dataframe with the addition of four win probability columns 
for the home and away
}
\description{
This function takes in the output of the game level play by play
function and returns the same dataframe with six new columns representing the
win probability for the home and away team pre and post snap and win probability
added.  The model used to calculate win probability is a genarlized addative
 model with three explanatory variables.
}

